rename to contentGatingHandler
