+++ packages/refresh-theme/start-server.js
@@ -17,7 +17,7 @@ const components = require('./components');
 const fragments = require('./fragments');
 const idxRouteTemplates = require('./templates/user');
 
-const defaultContentGating = () => false;
+const defaultContentGatingHandler = () => false;
 const routes = siteRoutes => (app) => {
   // Handle submissions on /__inquiry
   loadInquiry(app);
@@ -38,7 +38,7 @@ module.exports = (options = {}) => {
   const omedaConfig = getAsObject(options, 'siteConfig.omeda');
   const nativeXConfig = getAsObject(options, 'siteConfig.nativeX');
   const specGuideConfig = getAsObject(options, 'siteConfig.specGuides');
-  const contentGating = options.contentGating || defaultContentGating;
+  const contentGatingHanlder = options.contentGatingHandler || defaultContentGatingHandler;
   return startServer({
     ...options,
     routes: routes(options.routes),
@@ -46,7 +46,7 @@ module.exports = (options = {}) => {
     components: options.components || components,
     fragments: options.fragments || fragments,
     onStart: async (app) => {
-      set(app.locals, 'contentGating', contentGating);
+      set(app.locals, 'contentGatingHanlder', contentGatingHanlder);
       if (typeof onStart === 'function') await onStart(app);
       app.set('trust proxy', 'loopback, linklocal, uniquelocal');
       // Setup GAM.
