+++ packages/refresh-theme/templates/content/index.marko
@@ -11,7 +11,7 @@ $ const {
   GAM,
   req,
   userState,
-  contentGating
+  contentGatingHandler
 } = out.global;
 
 $ const showCompanySectionFilter = site.get("showCompanySectionFilters") || false;
@@ -181,7 +181,7 @@ $ const injectAds = ["article", "news", "press-release", "blog"].includes(type) 
               </if>
               <default-theme-content-contact-details obj=content />
 
-              $ const requiresRegistration = get(content, "userRegistration.isCurrentlyRequired") || contentGating(content);
+              $ const requiresRegistration = get(content, "userRegistration.isCurrentlyRequired") || contentGatingHandler(content);
               $ const accessLevels = getAsArray(content, "userRegistration.accessLevels");
               <marko-web-identity-x-access|context|
                 enabled=requiresRegistration
