+++ sites/sdcexec.com/index.js
@@ -4,7 +4,7 @@ const cufv1 = require('@ac-business-media/refresh-theme/redirect-handler-cufv1')
 const routes = require('./server/routes');
 const siteConfig = require('./config/site');
 const coreConfig = require('./config/core');
-const contentGating = require('./utils/content-gating');
+const contentGatingHandler = require('./utils/content-gating-handler');
 
 const { log } = console;
 
@@ -14,5 +14,5 @@ module.exports = startServer({
   siteConfig,
   routes,
   redirectHandler: cufv1('update.sdcexec.com'),
-  contentGating,
+  contentGatingHandler,
 }).then(() => log('Website started!')).catch(e => setImmediate(() => { throw e; }));
